chore: improve enrol course participants and tutorial members dialogs with scroll area
